Fix `instant-loan` test after changing API from `/holdings/3` to `/holdingsLogistics/v1`
I have updated the fixtures with fresh data to ensure that the data is as close to the reality as possible. Therefore, I have updated the test to align with the new fixture.

Since there was no instant loan for the material at the moment, I manually changed one of the branches. Because comments are not allowed in JSON, I added a key `dev_comment` with an explanation.

I also removed a non-helpful comment in the test.