Type `action` as `any`
Redux Toolkit v2 the action type changed from `any` to unknown, to encourage users to write type guards that checks the action object. The `action` object in `extractServiceBaseUrls()` was not originally written with types in mind, so we will keep it as `any` for now.
But it would be nice to introduce some type for this in the future.