Add `useGetWorkUrlFromIdentifier` to support material linking from loans
Since all data from loans comes from Publizon, we don’t have a `workId`, which is required to link to the material. Therefore, I use the `useComplexSearchWithPaginationQuery` to find the work using the ISBN (identifier) and construct the `workId` into a URL, e.g., `/work/work-of:870970-basis:62986115`.  Additionally, if it finds the material type, it adds that as a parameter.