Don't use link component for cases where we need to track, or other
This is a part of the movement to not use links ( <a/> ) when we need
to do more than just simply redirect. Due to us sometimes (this case
included) needing to do two separate things before redirecting, we in
the past had to hack our way to the <a/> tag not firing the redirect,
cuz we needed to first track the click + delete some persisted data.