Ensure that materials ready for loan display the correct "Lån før XX-XX-XXXX" date.
I am not a big fan of the `fetchDigitalMaterial` and `fetchMaterial` higher-order components. They are too difficult to read. Maybe these should be rewritten in the future. However, from what I can see, there is never an `identifier` in the props. Therefore, the `isDigital` variable is never set to true. This causes `getInfo` inside `src/apps/reservation-list/reservation-material/reservation-info.tsx` to use `pickupDeadline` instead of `expiryDate`, which is the intended behavior for digital materials.