Increase the limit of Facets being displayed to users to 50.
The intend of this change is to allow users to see more facets, which was limited to 5.

I've decided to start with changing this limit to 50 to begin with.

Option lists in a <select/> element, and are not easily styled. Because of this, i've also decided to see if the native renders of these will suffice before changing more markup.

DDFSAL-12