Only generate code coverage if any tests have been run
When running tests in parallel with Cypress Cloud orchestrating the
distribution we can end up in situations where a test run does not
actually run any tests. This in turn will not generate any code
coverage which again will make generating a code coverage report fail.
This in turn has made CI runs fail seemingly randomly.

To avoid this we refrain from running code coverage related actions
if no coverage information has been generated.

Our way to determine whether this is the case is based on the
existence of the nyc report directory. We currently use the default
value but now that we have a check for it outside of nyc we make the
value explicit to show the dependency.