Move MaterialGrid, Skeleton & utils to `/components/`folder.
I have separated all updates to other apps/components into other commits in order to not create an unreadable commit here.

For the component: This includes minor prop adjustments;

Make buttonText a prop being passed to the component,
Fixed issue with `initialMaximumDisplay`being locked to 4, this was never intended. It is now a prop defaulting to 4.

DDFSAL-3