Comment out generic response when updating patrons
For whatever reason our tool chain registers a generic response for
this request. This cannot be handled by other tools down the chain.

We do not care about the actual response here so it should be safe
to just comment it out even though it creates a more complex workflow
when updating the specification.