Fix deletion of data when change/pause reservation
When pausing the reservation, the `data` prop in `savePatron` only included the `onHold` data and did not include the email and phone number.

Therefore, I needed to add `...patron` to the `convertPatronSettingsV4toV6` function to ensure that all patron data is spread (including `onHold`), and that email address and phone nummer are properly converted.