Move facet mapping tracking into `addToFilter`
We should track clicks on terms from both `FacetLine` and `FacetBrowser` separately.

By introducing an `origin` property (`"facetLine" | "facetBrowser" | "facetUrl"`) on the `add` reducer and term, we can relocate all tracking logic into the `addToFilter` function within the `useFilterHandler` hook.

Here, we determine the origin of each interaction and utilize the updated `getAllFilterPathsAsString`, which now filters based on `origin`, ensuring accurate tracking data.

This ensures we track separately which terms are clicked in the `FacetBrowser` and the `FacetLine`.