Move `reservePhysicalButton` to page object and add `ensureReservationModalIsLoaded`
The `reservePhysicalButton` has been moved to the page object, as it belongs to the page itself and not the Reservation Modal.

Additionally, added `ensureReservationModalIsLoaded` to ensure that a material is selected before clicking the `submitButton`. This change improves test stability, as the `saveReservation` function in `ReservationModalBody` contains conditional logic that depends on `manifestationsToReserve`. Without waiting for this data to load, the function is  called with no effect, leading to flaky tests.