Initialize search query from URL parameter and improve autosuggest handling
This keeps the search string and ensures the dropdown with results is only open if the user has typed or rewritten the query and it is over 3 characters.

However, we now also support search by creator and subject, and this implementation does not account for those. So, if the user lands on the search page with a creator or subject search and rewrites it, it will perform a regular search with the modified string. The same issue applies to subject-based searches.