Allow any number <32 for material grid displays. DDFSAL-393
Originally, we only allowed a certain set of numbers to be displayed in
the material grid (4, 8, etc.), but DDF wants the editors to have to
choice to display odd-numbers - even if it means that the grid doesnt
have a clean end.

This removes the whole "calculateMaximumAmount" logic, and instead
relies on getting the length of the materials that gets sent along.

We still need a amount-input for the automatic grids, as they need a
limit.