Stop redirects to materials from suggestions with highlighted material
Suggestions from FBI have an odd data structure. They may relate to
topics or authors but still include a reference to a work. At the same
time we want to extract and display selected works as a part of the
suggestions.

In the current form these two types can get mixed up if a suggestion
is for an author or topic but include a reference to a work which is
also displayed directly. This will lead to the user being sent
directly to the work even though she selected the author/topic.

Avoid this by comparing the selected suggestion with the material
suggestions.

If the suggestion is a material suggestion when we can proceed as now.
Otherwise use the default approach by directing to the search page.

The helper isDisplayedAsWorkSuggestion got so simple and is only used
once so we might as well inline it.

DDFSAL-474