Improve filtering when selecting a suggested creator, topic or title
Currently we are not using the suggestion type for anything for the
top suggestion area in the search header. The selected term is merely
passed along as a query to the search page.

Improve on this by using them as filters where the suggestion type
matches a filter (creator or subject) or do a phrase search for titles
(as there is no title filter).

This should reduce the number of results returned and thus provide
more relevant results.

DDFSAL-474